Preparing release 0.9.1: warnings fix, Win CI build fix, build system improvement (#81)
* Fix build with C++17 enabled and system-provided boost package

* Fix build

* Fix build

* Build scripts cleanup and Readme file update

* Fix readme, improve appveyor  matrix

* Debug Win-based CI

* Debug Win-based CI

* Debug Win-based CI

* Fix runtime type bypass to gtest submodule

* Debug Win CI builds

* Debug Win CI builds

* Debug Win CI builds

* Debug Win CI builds

* Debug Win CI builds

* Debug Win CI builds

* Debug Win CI builds

* Debug Win CI builds

* Debug Win CI builds

* Debug Win CI builds

* Debug Win CI builds

* Debug Win CI builds

* Debug Win CI builds

* Debug Win CI builds

* Debug Win CI builds

* Debug Win CI builds

* Debug Win CI builds

* Debug Win CI builds

* Debug Win CI builds

* Debug Win CI builds

* Debug Win CI builds

* Debug Win CI builds

* Debug Win CI builds

* Debug Win CI builds

* Debug Win CI builds

* Debug Win CI builds

* Debug Win CI build
Fix MSVC warnings

* Debug Win CI build

* [skip ci] Update readme

+++ src/testers.cpp
@@ -215,7 +215,7 @@ bool ValueTester::Test(const InternalValue& baseVal, RenderContext& context)
         else if (valKind == ValueKind::Double)
         {
             auto dblVal = ConvertToDouble(val);
-            int64_t intVal = dblVal;
+            int64_t intVal = static_cast<int64_t>(dblVal);
             if (dblVal == intVal)
                 result = (testMode == (intVal & 1)) == (EvenTest ? 0 : 1);
         }
